move shared machinery from test_reactors.py to buildEmptyOperators.py
